+++ changelog.md
@@ -68,3 +68,4 @@
 * Introduced Neptune grass as a benthic character
 * Minor UI improvements
 * Randomised character movement and experimental prey movement
+* Improved collision detection